machineLearning/requirements.txt
fastapi==0.68.0
pandas==1.3.3
sqlalchemy==1.4.26
scikit-surprise==1.1.1
uvicorn==0.15.0
asyncio==3.4.3
python-decouple==3.5
fastapi[all]
python-dotenv